[Fix]検索機能のリファクタ&関数のprivate化&ファイル分けを実行
